everything working but javascript

+++ db/migrate/20150928073401_create_tickets.rb
@@ -2,10 +2,8 @@ class CreateTickets < ActiveRecord::Migration
   def change
     create_table :tickets do |t|
     	t.references :artwork, index: true
-    	t.integer :price
+    	t.float :price
     	t.integer :shares
-
-
       t.timestamps null: false
     end
     add_reference :tickets, :buyer, references: :users
